# Fix performance of `sparse_row(::ZZRing,::Int,::ZZRingElem)`