Before:
```julia-repl
julia> R=ZZ; v=R(1);

julia> @btime sparse_row($R, [(1,$v)])
  124.772 ns (9 allocations: 416 bytes)
Sparse row with positions [1] and values ZZRingElem[1]

julia> @btime sparse_row($R, [1],[$v])
  107.479 ns (11 allocations: 464 bytes)
Sparse row with positions [1] and values ZZRingElem[1]

julia> @btime sparse_row($R, 1, $v)
  178.194 ns (14 allocations: 528 bytes)
Sparse row with positions [1] and values ZZRingElem[1]

julia> @btime sparse_row($R, 1, $v; check=false)
  175.253 ns (14 allocations: 528 bytes)
Sparse row with positions [1] and values ZZRingElem[1]
```

After:
```julia-repl
julia> @btime sparse_row($R, 1, $v)
  96.549 ns (7 allocations: 208 bytes)
Sparse row with positions [1] and values ZZRingElem[1]

julia> @btime sparse_row($R, 1, $v; check=false)
  96.796 ns (7 allocations: 208 bytes)
Sparse row with positions [1] and values ZZRingElem[1]
```

CC @HechtiDerLachs 

Resolves https://github.com/oscar-system/Oscar.jl/issues/5371 (well, the last bit, the bulk was done by @HechtiDerLachs thanks)